backend/SurplusApp/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart
fastapi-mail
python-dotenv
cloudinary
pillow
apscheduler
alembic